+++ lib/email_engine/engine.rb
@@ -1,0 +1,8 @@
+module EmailEngine
+  class Engine < ::Rails::Engine
+    initializer "email_engine" do |app|
+      secrets = app.respond_to?(:secrets) ? app.secrets : app.config
+      EmailEngine.secret_token ||= secrets.respond_to?(:secret_key_base) ? secrets.secret_key_base : secrets.secret_token
+    end
+  end
+end