+++ CHANGELOG.md
@@ -44,7 +44,7 @@
 
 ## 0.1.2
 
-- Added `AhoyEmail.track` (fix)
+- Added `EmailEngine.track` (fix)
 
 ## 0.1.1